requirements.txt
# Web Framework
flask==2.3.3
gunicorn==21.2.0
Werkzeug==2.3.7

# REST API Extensions
flask-restful==0.3.10
flask-cors==4.0.0

# Machine Learning
scikit-learn==1.3.2
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4
joblib==1.3.2
matplotlib==3.8.2

# Data Processing
python-dateutil==2.8.2
pytz==2023.3

# Metrics & Monitoring
prometheus-flask-exporter==0.22.4

# Utilities
requests==2.31.0
python-dotenv==1.0.0

# Model Serialization
cloudpickle==2.2.1

tensorflow==2.19.0  # เวอร์ชันล่าสุดที่รองรับ Python 3.11
torch==2.1.2
transformers==4.36.2
keras==3.9.0

opencv-python-headless==4.8.1.78
pillow==10.2.0
albumentations==1.3.1